Fix #27. (#28)
The issue was in Schoenhage reduction. We compute a=<b0,b0> first in case we have to exit early, but we mistakenly recompute it if no early exit happens. Since this value is computed by accumulating into a GMP integer, this incorrectly results in a=2*<b0,b0>, and the resulting Schoenhage computation fails.